backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mlh-pe-hackathon"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "faker>=33.0",
    "flask>=3.1",
    "gunicorn>=23.0",
    "peewee>=3.17,<4.0",
    "psycopg2-binary>=2.9",
    "python-dotenv>=1.0",
]

[tool.setuptools.packages.find]
include = ["app*"]
